Ignore generated .xsl files.
Without this, cvs update reports all generated .xsl files in this
directory as unknown. It's a big list, and it'd be a lot easier to
read the cvs update output if these files weren't listed every time.

Note: This does not cause CVS to ignore the html2xhtml.xsl file or
to treat it in any way way differently than it would if the *.xsl
line in this .cvsignore file did not exist. It doesn't seem to at
least. If that's wrong, this change can always be reverted.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@3790 6cfa91a5-b114-0410-aa58-ce7c4e7541e0