Generate XEP bookmarks for Refentry children. (closes #1212491)
Titled child sections of Refentry (Refsynosisdiv, Refsection, and
Refsect1 to Refsect3) were not included in the match statement
used for generating XEP bookmarks. Don't know whether that was
intentional for some reason or whether it was just an oversight.
But given that both AXF and Passivetex bookmarks are generated for
those same Refentry children, it seems like XEP ones ought to also
be generated, for the sake of consistency if for no other reason.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@4881 6cfa91a5-b114-0410-aa58-ce7c4e7541e0