Align refnamediv title correctly when refentry.generate.title is non-zero (closes #1212641).
::Problem:
When refentry.generate.title is non-zero, the title output for
Refnamediv is not aligned flush left.

::Cause:
No code for setting start-indent="" was included in
refentry.title.properties. It should be in order to make the
Refnamediv title output be flush left, as are titles for all other
sectioning children of Refentry

::Fix:
Added code for setting start-indent="" in
refentry.title.properties.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@4882 6cfa91a5-b114-0410-aa58-ce7c4e7541e0