Don't render NAME heading for secondary Refnamedivs (closes #1216292)
If a document has multiple Refnamedivs, a NAME heading was getting
rendered for each. But we only need one NAME heading. This change
causes it to be rendered just once. This makes behavior in this
respect consistent with how the HTML and FO stylesheets handle the
generated NAME heading for Refnamediv.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@4895 6cfa91a5-b114-0410-aa58-ce7c4e7541e0