Reworked *info gathering and rethought Refclass handling.
For each Refentry found, we now cache its *info and its parent's
*info as node-sets; we then do all further matches against those
node-sets (rather than re-selecting the original *info nodes each
time we need to check them).

Also, reverted the special handling of Refclass that was added
recently. We eventually need to make Refclass handling consistent
with that of the HTML and FO stylesheets.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@4896 6cfa91a5-b114-0410-aa58-ce7c4e7541e0