First shot at implementing table support in the manpages stylesheet. This works well as-is for normal span-less tables. It even works for tables with colspans. But it does not yet even attempt to do anything for rowspans. It will take more work to add rowspan handling that. Not sure if it's worth the work, really.
Note: This table support is implemented by do a first pass to
process each table using the templates in the HTML stylesheet
(which we import). We don't override any of the <row> or <tr> or
<entry> or <td>, etc., table-markup templates, but we do override
the block and inline templates for whatever the child content of
the <entry> and <td> instances. So we end up with a node that has
"hybrid" output -- with table rows and cells marked up with HTML
<tr> and <td> tags, with with the <td> child content marked up in
roff.

So we use exsl:node-set on that node and make a second-pass
through all of it to replace the <tr> and <td> parts with tbl(1)
macros. Voila -- the end result is properly marked up tbl/roff
output for the table.

One nice thing about doing it this is that it supports both CALS
and HTML table markup, without needing any conditional logic to
deal with CALS separately. Because the first pass causes both the
CALS and HTML source markup to be "normalized" into HTML.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@5325 6cfa91a5-b114-0410-aa58-ce7c4e7541e0