In manpages output, emit warnings for notesources (footnote, etc.) that have something other than para as a child.
The numbered-with-hanging-indent formatting that's used for
rendering endnotes in the NOTES section of man pages places some
limits/assumptions on how the DocBook source is marked up; namely,
for notesources (footnote, annotation, etc.) that can contain
block-level children, if the they have a block-level child such as
a table or itemizedlist or orderedlist that is the first child of
a footnote, we have no way of rendering/indenting its content
properly in the endnotes list.

Thus, the manpages stylesheet not emits a warning message for that
case, and suggests the "fix" (which is to wrap the table or
itemizedlist or whatever in a para that has some preferatory text.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@6840 6cfa91a5-b114-0410-aa58-ce7c4e7541e0