Changed handling of bold and italic/underline output in manpages output. Should be transparent to users, but...
This touches handling of all bold and italic/underline output. The
exact change is that the mode="bold" and mode="italic" utility
templates were changed to named templates. (I think maybe I've
changed it back and forth from mode to named before, so this is
maybe re-reverting it yet again).

Anyway, the reason for the change is that the templates are
sometimes call on dynamically node-sets, and using modes to format
those doesn't allow passing info about the current/real context
node from the source (not the node-set created by the stylesheet)
to that formatting stage.

The named templates allow the context to be passed in as a
parameter, so that the bold/ital formatting template can use
context-aware condition checking.

This was basically necessary in order to suppress bold formatting
in titles, which otherwise gets screwed up because of the numbnut
way that roff handles nested bold/ital.

Closes #1674534). Much thanks to Daniel Leidert, whose in his
docbook-xsl bug-finding kung-fu has achieved Grand Master status.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@6843 6cfa91a5-b114-0410-aa58-ce7c4e7541e0