Added to default charmap used by manpages:
  - the "letters" part of the 'C1 Controls And Latin-1 Supplement
    (Latin-1 Supplement)' Unicode block
  - Latin Extended-A block (but not all of the characters from
    that block have mappings in groff, so some of them are still
    passed through as-is)

The effects of this change are that in man pages generated for
most Western European languages and for Finnish, all characters
not part of the Roman alphabet are (e.g., "accented" characters)
are converted to groff escapes.

Previously, by default we passed through those characters as is
(and users needed to use the full charmap if they wanted to have
those characters converted).

As a result of this change, man pages generated for Western
European languages will be viewable in some environments in which
they are not viewable if the "raw" non-Roman characters are in them.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@7734 6cfa91a5-b114-0410-aa58-ce7c4e7541e0