Changed handling of hanging indents for cmdsynopsis, funcsynopsis, and synopfragment such that they now look correct in non-TTY/PS output. We now use the groff \w escape to hang by the actual width -- in the current font -- of the command, funcdef, or synopfragment references number (as opposed to hanging by the number of characters). This rendering in TTY output remains the same, since the width in monospaced TTY output is the same as the number of characters.
Also, created new synopsis-block-start and synopsis-block-end
templates to use for cmdsynopsis and funcsynopsis instead of the
corresponding verbatim-* templates.

Along with those changes, also corrected a problem that caused the
content of synopfragment to be dropped, and made a
vertical-spacing change to adjust spacing around table titles and
among sibling synopfragment instances.


git-svn-id: https://svn.code.sf.net/p/docbook/code/trunk/xsl@7956 6cfa91a5-b114-0410-aa58-ce7c4e7541e0